Inventory Feedback
Added movement to UI items when selection changes. Instead of teleporting left and right they lerp now and it gives such a better feel to swapping weapons. Also added a small shrink on the arrows left and right. All adds to a good feeling swapping active item left and right.